transfer completion callback called on flush
The transfer completion callback is called with size = -1 when a
transfer is canceled.

This makes it easier for the caller to keep track of its state,
e.g. how many transfers are in progress. In most cases, transfers are
not flushed without the user knowing, but when hotplugging a USB device,
the endpoints (and thus the transfers) are flushed on re-enumeration.

In the future the library could have other cases (error handling?) where
it is necessary to flush an endpoint + transfers.